chore(curve): refactor pools, tokens, and contracts structure
