+++ src/entities/trades/curve/contracts/contracts.ts
@@ -1,23 +1,32 @@
 import { JsonRpcProvider } from '@ethersproject/providers'
-import { BigNumber, BigNumberish } from '@ethersproject/bignumber'
 import { Contract } from '@ethersproject/contracts'
-
-import { ZERO_ADDRESS } from '../../../constants'
-import { ChainId } from '../../../constants'
-import { TOKENS_MAINNET } from './constants'
+import { ZERO_ADDRESS } from '../../../../constants'
+import { ChainId } from '../../../../constants'
+import { TOKENS_MAINNET } from '../tokens'
 // ABIs: trimmed for bundle size
-import { ADDRESS_PROVIDER_ABI, CURVE_ROUTER_ABI, REGISTRY_EXCHANGE_ABI } from './abi'
+import {
+  REGISTRY_EXCHANGE as REGISTRY_EXCHANGE_ABI,
+  ADDRESS_PROVIDER as ADDRESS_PROVIDER_ABI,
+  CURVE_ROUTER as CURVE_ROUTER_ABI,
+} from '../abi'
+
+import {
+  GetBestPoolAndOutputParams,
+  GetBestPoolAndOutputResult,
+  GetExchangeRoutingInfoParams,
+  GetExchangeRoutingInfoResults,
+} from './types'
 
 // Constants
 export const MAINNET_CONTRACTS = {
   addressProvider: '0x0000000022d53366457f9d5e68ec105046fc4383',
-  router: '0xfA9a30350048B2BF66865ee20363067c66f67e58'
+  router: '0xfA9a30350048B2BF66865ee20363067c66f67e58',
 } as const
 
 export const RPC_PROVIDER_LIST = {
   [ChainId.MAINNET as ChainId]: 'https://mainnet.infura.io/v3/e1a3bfc40093494ca4f36b286ab36f2d',
   [ChainId.XDAI as ChainId]: 'https://rpc.xdaichain.com/',
-  [ChainId.ARBITRUM_ONE as ChainId]: 'https://arb1.arbitrum.io/rpc'
+  [ChainId.ARBITRUM_ONE as ChainId]: 'https://arb1.arbitrum.io/rpc',
 }
 
 /**
@@ -26,26 +35,6 @@ export const RPC_PROVIDER_LIST = {
 export const getProvider = (chainId: ChainId) => {
   const host = RPC_PROVIDER_LIST[chainId]
   return new JsonRpcProvider(host)
-}
-
-export interface GetBestPoolAndOutputParams {
-  tokenInAddress: string
-  tokenOutAddress: string
-  amountIn: BigNumberish
-  chainId: ChainId
-}
-export type GetExchangeRoutingInfoParams = GetBestPoolAndOutputParams
-
-export interface GetBestPoolAndOutputResult {
-  expectedAmountOut: BigNumber
-  poolAddress: string
-  registryExchangeAddress: string
-}
-
-export interface GetExchangeRoutingInfoResults {
-  routes: string[]
-  indices: BigNumber[]
-  expectedAmountOut: BigNumber
 }
 
 /**
@@ -57,7 +46,7 @@ export async function getBestCurvePoolAndOutput({
   amountIn,
   tokenInAddress,
   tokenOutAddress,
-  chainId
+  chainId,
 }: GetBestPoolAndOutputParams): Promise<GetBestPoolAndOutputResult | undefined> {
   if (chainId !== ChainId.MAINNET) {
     throw new Error('Best Pool Find is only available on Mainnet')
@@ -73,7 +62,7 @@ export async function getBestCurvePoolAndOutput({
   const tricryptoCoins = [
     TOKENS_MAINNET.usdt.address.toLowerCase(),
     TOKENS_MAINNET.wbtc.address.toLowerCase(),
-    TOKENS_MAINNET.weth.address.toLowerCase()
+    TOKENS_MAINNET.weth.address.toLowerCase(),
   ]
 
   if (tricryptoCoins.includes(tokenInAddress.toLowerCase()) && tricryptoCoins.includes(tokenOutAddress.toLowerCase())) {
@@ -81,7 +70,7 @@ export async function getBestCurvePoolAndOutput({
   }
 
   const registryExchangeAddress = await addressProviderContract.get_address(2, {
-    gasLimit: 100000 // due to Berlin upgrade. See https://github.com/ethers-io/ethers.js/issues/1474
+    gasLimit: 100000, // due to Berlin upgrade. See https://github.com/ethers-io/ethers.js/issues/1474
   })
 
   const registryExchangeContract = new Contract(registryExchangeAddress, REGISTRY_EXCHANGE_ABI, getProvider(chainId))
@@ -98,7 +87,7 @@ export async function getBestCurvePoolAndOutput({
   return {
     poolAddress,
     expectedAmountOut,
-    registryExchangeAddress
+    registryExchangeAddress,
   }
 }
 
@@ -117,7 +106,7 @@ export function getRouter() {
 export async function getExchangeRoutingInfo({
   amountIn,
   tokenInAddress,
-  tokenOutAddress
+  tokenOutAddress,
 }: GetExchangeRoutingInfoParams): Promise<GetExchangeRoutingInfoResults | undefined> {
   const routerContract = getRouter()
 
@@ -125,7 +114,7 @@ export async function getExchangeRoutingInfo({
     const params = [tokenInAddress, tokenOutAddress, amountIn.toString()]
 
     const exchangeRoutingRes = await routerContract.get_exchange_routing(...params, {
-      from: ZERO_ADDRESS
+      from: ZERO_ADDRESS,
     })
 
     const [routes, indices, expectedAmountOut] = exchangeRoutingRes
@@ -133,7 +122,7 @@ export async function getExchangeRoutingInfo({
     return {
       expectedAmountOut,
       indices,
-      routes
+      routes,
     }
   } catch (error) {
     // Throw any non-EVM errors
